Don't use `asyncio.run`
While nice in theory, `asyncio.run` handles shutdown in a way that we
don't necessarily want. When an exception is raised in the loop (like a
`SystemExit`), all running tasks are cancelled at the same time.
Instead, we'd like to cancel the *main* task and let it handle cleaning
up child tasks. We then check if any tasks remain, cancel and log those
(lingering tasks indicate bugs/failures), and clean up the loop.
